[Inference API] Extract AmazonBedrock types into a dedicated file
Move definitions to AmazonBedrockTypes.ts; trim CommonTypes.ts.

Made-with: Cursor
